LogoutTest class added
